finalized the ml email classifier
Made a jupyer notebook that took an initial sample and trained a model and iteratively improived it by labeling data manually unitl eventually the model was confident enough to predict. Ended with an accuracy of 98%. Have a relatively balanced dataset. Used 3 different datasets and combined them. Used active learning to label efficiently.